pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "akernel"
version = "0.2.1"
description = "An asynchronous Python Jupyter kernel"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "David Brochart", email = "david.brochart@gmail.com"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
keywords = [ "jupyter" ]
dependencies = [
    "python-dateutil",
    "colorama",
    "gast >=0.6.0, <0.7.0",
    "comm >=0.1.3,<1",
]

[dependency-groups]
test = [
    "mypy",
    "ruff",
    "pytest",
    "pytest-asyncio",
    "pytest-rerunfailures",
    "types-python-dateutil",
    "kernel_driver >=0.0.7",
    "ipyx >=0.1.7",
    "zict",
]

[project.optional-dependencies]
subprocess = [
    "zmq-anyio >=0.3.9,<0.4.0",
    "typer >=0.4.0",
]

react = [
    "ipyx >=0.1.7",
]

cache = [
    "zict",
]

[project.scripts]
akernel = "akernel.akernel:cli"

[tool.hatch.build.targets.wheel]
ignore-vcs = true
packages = ["src/akernel"]

[tool.hatch.build.targets.wheel.shared-data]
"share/jupyter/kernels/akernel/kernel.json" = "share/jupyter/kernels/akernel/kernel.json"

[project.urls]
Homepage = "https://github.com/davidbrochart/akernel"

[tool.ruff]
line-length = 100
exclude = ["examples"]

[tool.uv.sources]
fps-akernel-task = { workspace = true }
